requirements.txt
flask==3.1.1
flask-socketio==5.5.1
python-socketio==5.13.0
pynput==1.8.1
Pillow==11.2.1
qrcode==8.1
pystray==0.19.5
pyinstaller==6.12.0
